units: systemd-logind fails hard without dbus
That is, without --enable-kdbus and kdbus running.

With --enable-kdbus things are more complicated, because dbus might be
necessary, if kdbus is missing at runtime. If it is not necessary,
the socket will be started, which is not imporant, but not the service.

(cherry picked from commit 8f9c6fe5ff1d59001aecbf3fbf9ca0ed7ff28ba7)