getty-generator: verify ttys before we make use of them
The ttyS[0-3] devices are weird. They may be enumerated, but when one
actually tries to open and use them they return EIO, because they don't
actually exist. Because they may be enumerated they may be specified on
the kernel command line as console=. And some people do that as default.
As response to that we'll spawn a getty on the tty that will quickly
fail, and we retry a couple of time before giving up. That is quite
noisy.

With this new change we will validate all serial terminals configured
with console= on the kernel cmdline before adding gettys on them, and
remove the invalid ones. THis should remove the noise later on.

This should make Eric Paris happy!

(cherry picked from commit 543407517e05688915128cfc544c5d7f97f240ef)