core: print debug instead of error message
In case that session scope is in some other state than SCOPE_RUNNING or
SCOPE_ABANDONED we will not print error message. It might be the case
that scope is in other state because systemd realized that scope is
empty before dispatching dbus call from pam hook. Or system is going
down and stop job is runnig for the scope and changed scope state to
SCOPE_STOP_SIGTERM or SCOPE_STOP_SIGKILL.