Manager: Fix daemon-reload crash Manager: Parse Slice= even without CGroups
On daemon-reload, we didn't clear the PTGroups hashmap, so old
PTGroup-to-Unit mappings remained, which were now invalid. We also
didn't get rid of the old PTGroups. Let us instead release all
PTGroups and clear that hashmap.

Slice= config file directive was hidden under an M4 conditional on
Use_CGroups. We have slices even for non-CGroups systems so move
them out of the conditional.