shared: Process >1 kevents in dbus-loop-kqueue
The PAM module was mysteriously hanging waiting for a reply on
some logins in NetBSD. Then I noticed that if I ran 'iwloginctl'
elsewhere while the PAM module was hanging, that seemed to unhang
the PAM module.

The dbus-loop-kqueue is integrated into logind's main event loop
(E-Poll based) by watching the kqueue FD itself; it appears the
KQueue triggers an event when it has new events ready and hasn't
yet been called to process any of them; but having processed just
one of perhaps several events that were queued at once, it does not
trigger another event so as to make the main E-Poll loop return.

So now we loop calling kqueue() until it has 0 events left for us.
I think this is adequate. I believe we do something similar,
nesting a KQueue inside an E-Poll outer loop, in the service
manager, so we should investigate that too, lest we lose process
events!

I also added an 8s timeout to the logind message sent by the PAM
module - I think the default D-Bus timeout of 25s is too long to
wait for a login.