Try to track subprocesses on macOS
We do similarly to launchd and iterate over every child PID of the
process we're watching on receiving NOTE_FORK. First all are added
to the kqueue in one loop, then a second loop puts them in the
PTGroups structures. We do it thus to help us to race against the
possibility of further subprocesses being created or processes
exiting and other TOCTTOU problems.